Major refactor of Debugger.java
The Debugger class was getting too big and was doing
too much so it was time to break it down. We seperated
various roles into individual classes. Roles including:
requesting events, polling for events and
evaluating the stack frame. We also introduce
the EventObserving interface that allow
observers to subscribe to debugger events
eg. onBreakpointEvent, onStepEvent, etc..